📱 Improve Responsiveness
