+++ next/src/components/landing/Hero.tsx
@@ -13,11 +13,11 @@ import SparkleIcon from "../../../public/icons/sparkle-default-regular.svg";
 import GlowWrapper from "../GlowWrapper";
 import Spline from "@splinetool/react-spline";
 
-const Hero: React.FC = () => {
+const Hero: React.FC<{ className?: string }> = ({ className }) => {
   const router = useRouter();
 
   return (
-    <div className="relative grid h-screen grid-cols-1 place-items-center gap-2 lg:grid-cols-2">
+    <div className={clsx("grid grid-cols-1 place-items-center gap-2 lg:grid-cols-2", className)}>
       <div className="z-10 col-span-1">
         <FadeIn duration={3} initialY={50} className="flex flex-col gap-12">
           <BannerBadge href="https://calendly.com/reworkdai/enterprise-customers" target="_blank">
